+++ mvc/src/test/java/com/interface21/webmvc/servlet/view/JsonViewTest.java
@@ -1,7 +1,6 @@
 package com.interface21.webmvc.servlet.view;
 
 import com.interface21.web.http.MediaType;
-import com.interface21.webmvc.servlet.View;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.DisplayName;
